chore(core): rename style module to styles
Plural feels more appropriate since this is doing a number of different styling operations.
